Make lightstep-tracer tests pass
Many tests were deleted, but that is a good thing: they were testing
things that basictracer-python does for lightstep-tracer-python now.

There were some warts with the way this test infrastructure was set up,
and I attempted to address those here.

One functional change was that reports are no longer sent when there's
no buffered span data.
