+++ tests/opentracing_compatibility_test.py
@@ -5,12 +5,6 @@ from opentracing.harness.api_check import APICompatibilityCheckMixin
 
 import lightstep.tracer
 
-# The binary carrier is not yet implemented in LightStep. Disable this OpenTracing
-# test until it is (otherwise it makes it far more difficult to detect legitimate
-# regressions in the unit tests while this known failure exists).
-def test_alway_pass(self):
-    return
-APICompatibilityCheckMixin.test_binary_propagation = test_alway_pass
 
 class LightStepTracerOpenTracingCompatibility(unittest.TestCase, APICompatibilityCheckMixin):
     """This unittest currently passes, but then the "Flush Thread" drops a stack trace.
@@ -20,9 +14,11 @@ class LightStepTracerOpenTracingCompatibility(unittest.TestCase, APICompatibilit
     def tracer(self):
         return lightstep.tracer.init_tracer(periodic_flush_seconds=0)
 
+
 class DebugTracerOpenTracingCompatibility(unittest.TestCase, APICompatibilityCheckMixin):
     def tracer(self):
         return lightstep.tracer.init_debug_tracer()
+
 
 if __name__ == '__main__':
     unittest.main()
