+++ tox.ini
@@ -1,15 +1,11 @@
 [tox]
-envlist = py26, py27
-
 ; commented out py34 testing due to current thrift issues with python3
 ; envlist = py26, py27, py34
+envlist = py26, py27
+
 
 [testenv]
 commands =
-    python tests/runtime_test.py
-    python tests/util_test.py
     python tests/opentracing_compatibility_test.py
-deps =
-    opentracing==1.0rc3
-    jsonpickle==0.9.2
-    thrift==0.9.2
+    python tests/recorder_test.py
+    python tests/util_test.py
